Deprecate class `CompatDecorator`
Didn't notice this at first when implementing `IcingaFormDecorator`.
Though now I don't see a reason for it anymore. It's by far less
capable and resides in the compat namespace. Do we have a replacement
in the ipl for Icinga Web 2? (routing, sessions, auth) I don't know
of any, and even if planned, this is so long way off that calling
this a compatilibity implementation is a bit fast.