CompatController: Allow to mix old-style actions with new-style ones
This will now only render the controls and footer if there's any content.
This is based on the assumption that an action has to define explicit
content if it wants to show controls or a footer. Even if the content
is populated by Javascript, there's usually another container part of
the content and the content is not directly utilized as target.