AnnuallyFields: Drop usage of `onRegistered()` handler
The `onRegistered()` mechanism works with `FieldsetElement` only when
the fieldset is being used at the first level, i.e if the fieldset is
added directly to the concrete `Form`. Otherwise, with such a nested
`FieldsetElement` as the `AnnuallyFields`, this mechanism doesn't work
as it is supposed to. Howeever, due to all the validation improvements
made in `ipl-/html,validator,web`, this element doesn't even need to use
that handler to work, so it's dropped.