SortControl: Make the constructor private
A future update will remove support for the $url parameter
and the columns are mandatory anyway. And since we use late
static binding in self::create(), we can guarantee that the
instance creation will always succeed.

Currently, no released code uses the constructor directly,
so it's safe to adjust the visiblity.