ActionList: remove click selector `not(.page-selector)`
it somehow became irrelevant, but must have been important before. But now that we don't want any icingaweb-specific code here, it has to be removed anyway.

it become somehow inrelevent, must be relevant before. But now as we dont want any icingaweb specific code here it must be removed anyway.