Introduce new item list/header layout implementation
Represents mostly what was already part of Icinga DB Web
in the form of view modes. The important change is that
what was previously list layout, is now just item layout.

Allowing the re-use of the same implementation of any given
layout in a list or list-like container, but also in any other,
e.g. a detail header.

Just implement a renderer for a particular item type and couple
it with the desired layout. Add the result to your list item,
header, flyout, whatever and… you're done!