TermInput: Keep original minimum width of terms
The remove label is now a title which allows terms
to be as narrow as possible again.

justify-content which was previously used to center
the label's contents cannot be used anymore though.
It has the effect of pushing overflowing content
also to the left, which causes the icon to not be
centered at the minimum width. Plain old, but working,
auto margins are the solution here.

fixes #256