CsrfCounterMeasure: Allow external control
Allows to set the id using a setter, avoiding the
need to access the session inside a form for example.
A controller can now set the id. The same goes for
deciding whether a token is required or not, which
is necessary in unit tests, where it is not.

`createCsrfCounterMeasure()` has now been deprecated
in favor of `addCsrfCounterMeasure()` since no element
can be returned if disabled and 99% of all usages
immediately add the *created* element to the form anyway.